add doc build process, and supporting code changes
- hide timeticks_ms from sphinx:

time.ticks_ms is a micropython specific function and isn't available at documentation generation time

- add publish docs workflow
- Add the ability to detect a test env

The TEST_ENV envvar can be set to indicate that we are in a test
environment and have no hardware. For example when generating
documentation or running automated tests.

Note that os.getenv() does not exist _yet_ in micropython v1.18. It looks like it
will be available in the next version. Instead we'll try to import
micropython in order to determine our environment.

- remove a docgen warning and make the code reference clearer
- treat doc warnings as errors
- add a root level Makefile, delegating doc related tasks to the docs/Makefile