move the display of the bootspash to the menu
we don't want to run any code on europi import that we don't have to
because it can break things. In this case the docs build.

This has advantage of only showing the bootsplash when it is really
needed (during the slow menu load) and not when it isn't (during a fast
single script load).