LockableKnob init fix (#193)
* add back in the KnobPlayground

* change var name in test

* add initial_percentage_value and initial_uint16_value to LockableKnob

This makes the values used in initial knob setup more clear.

* update TuringMachine for LockableKnob changes

add ability to set the initial value in the TM class. Use this to more clearly setup the inital state of the locked knob.

* remove KnobPlayground from menu

We'd like to make this available, but there seems to be a memory
allocation issue when adding more scripts. This will be addressed in a
different PR.